% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Delivers praise to the user}
\usage{
praise(name = "Vis")
}
\arguments{
\item{name}{Name of the person you want to praise. Must be a character. Default's to the name, "Vis".}
}
\value{
a nice message to user
}
\description{
This is an example piece of documentation.
}
\examples{
praise(name = "Vis")
}
